Move Header to GhBase for 625XT/505/580

+++ trunk/Device/Gb580Device.cs
@@ -64,7 +64,7 @@ namespace ZoneFiveSoftware.SportTracks.Device.Globalsat
             foreach (Gb580Packet.TrackFileHeader header in headers)
             {
                 totalPoints += header.TrackPointCount;
-                trackIndexes.Add(header.TrackPointIndex);
+                trackIndexes.Add((Int16)header.TrackPointIndex);
             }
             float pointsRead = 0;
